Func AddOrder (Doing)

+++ BMA/BMA/Models/Product.cs
@@ -28,7 +28,7 @@ namespace BMA.Models
         public string Descriptions { get; set; }
         public string Note { get; set; }
         public string ProductImgage { get; set; }
-        public double ProductStandardPrice { get; set; }
+        public int ProductStandardPrice { get; set; }
         public Nullable<int> CategoryId { get; set; }
         public bool IsActive { get; set; }
         public string ProductCode { get; set; }
